requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
marshmallow
flask-jwt-extended
passlib
flask-migrate